Extract the ACF and PACF parameters of time series and their model residuals

This function is included in ts.analysis function and aims to extract the ACF and PACF details of the input time series data and the ACF, PACF of the residuals after fitting an Arima model.

ts.acf(tsdata, model_residuals, a=0.95, tojson=F)

Arguments

tsdata
The input univariate time series data

model_residuals
The model's residuals after fitting a model to the time series

a
The significant level (default a=0.95)

tojson
If TRUE the results are returned in json format, default returns a list

Value

A list with the parameters:

acf.parameters:

acf The estimated acf values of the input time series

acf.lag The lags at which the acf is estimated

confidence.interval.up The upper limit of the confidence interval

confidence.interval.low The lower limit of the confidence interval

pacf.parameters:

pacf The estimated pacf values of the input time series

pacf.lag The lags at which the pacf is estimated

confidence.interval.up The upper limit of the confidence interval

confidence.interval.low The lower limit of the confidence interval

acf.residuals.parameters:

acf.res The estimated acf values of the model residuals

acf.res.lag The lags at which the acf is estimated of the model residuals

confidence.interval.up The upper limit of the confidence interval

confidence.interval.low The lower limit of the confidence interval

pacf.residuals.parameters:

pacf.res The estimated pacf values of the model residuals

pacf.res.lag The lags at which the pacf is estimated of the model residuals

confidence.interval.up The upper limit of the confidence interval

confidence.interval.low The lower limit of the confidence interval

Details

This function is used internally in ts.analysis function and the output is a list with grouped ACF and PACF parameters of the input time series data, as well as the ACF and PACF parameters of the residuals needed for the graphical purposes in OBEU.

See also

ts.analysis, Acf, Pacf

Examples

ts.acf(Athens_draft_ts)

#> $acf.parameters
#> $acf.parameters$acf
#>  [1]  1.00000000  0.67010083  0.20480616 -0.29189364 -0.51269294 -0.51430701
#>  [7] -0.28195630 -0.01821677  0.10932753  0.09101177  0.03254134
#> 
#> $acf.parameters$acf.lag
#>  [1]  0  1  2  3  4  5  6  7  8  9 10
#> 
#> $acf.parameters$confidence.interval.up
#> [1] 0.5657929
#> 
#> $acf.parameters$confidence.interval.low
#> [1] -0.5657929
#> 
#> 
#> $pacf.parameters
#> $pacf.parameters$pacf
#>  [1]  0.67010083 -0.44327501 -0.43580079  0.02709037 -0.11788115  0.01627274
#>  [7] -0.04326813 -0.27712013 -0.12208959  0.05034829
#> 
#> $pacf.parameters$pacf.lag
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $pacf.parameters$confidence.interval.up
#> [1] 0.5657929
#> 
#> $pacf.parameters$confidence.interval.low
#> [1] -0.5657929
#> 
#> 
#> $acf.residuals.parameters
#> NULL
#> 
#> $pacf.residuals.parameters
#> NULL
#>