Model fit of non seasonal time series

Model fit of non seasonal time series

ts.non.seas.model(tsdata, x.ord=NULL, tojson=F)

Arguments

tsdata
The input univariate non seasonal time series data

x.ord
An integer vector of length 3 specifying the order of the Arima model

tojson
If TRUE the results are returned in json format, default returns a list

Value

A list with the following components:

model.summary:

ts_model The summary model details returned as Arima object for internal use in ts.analysis function

model:

ts_model:

arima.order The Arima order

arima.coef A vector of AR, MA and regression coefficients

arima.coef.se The standard error of the coefficients

residuals: The residuals of the model (fitted innovations)

compare:

variance.coef The matrix of the estimated variance of the coefficients

resid.variance The MLE of the innovations variance

not.used.obs The number of not used observations for the fitting

used.obs the number of used observations for the fitting

loglik The maximized log-likelihood (of the differenced data), or the approximation to it used

aic The AIC value corresponding to the log-likelihood

bic The BIC value corresponding to the log-likelihood

aicc The second-order Akaike Information Criterion corresponding to the log-likelihood

Details

Model fit of non seasonal time series using arima models of non seasonal time series data. The model with the lowest AIC value is selected for forecasts.

See also

ts.analysis, Arima

Examples

ts.non.seas.model(Athens_draft_ts)

#> $model.summary
#> Series: tsdata 
#> ARIMA(1,1,1) 
#> 
#> Coefficients:
#>          ar1     ma1
#>       0.1441  0.0153
#> s.e.  0.8665  0.8392
#> 
#> sigma^2 estimated as 1.107e+16:  log likelihood=-217.7
#> AIC=441.41   AICc=444.84   BIC=442.6
#> 
#> $model
#> $model$arima.order
#> [1] 1 1 0 0 1 1 0
#> 
#> $model$arima.coef
#>        ar1        ma1 
#> 0.14413158 0.01530798 
#> 
#> $model$arima.coef.se
#>    ar1    ma1 
#> 0.8665 0.8392 
#> 
#> 
#> $residuals_fitted
#> $residuals_fitted$residuals
#> Time Series:
#> Start = 2004 
#> End = 2015 
#> Frequency = 1 
#>  [1]     720894.6  -90787238.6    4239112.9  107712202.6  117144839.4
#>  [6]   39442974.7   48646734.4  -54982432.3  -57635036.0 -189145649.1
#> [11]  137951183.7  -31298309.8
#> 
#> $residuals_fitted$fitted
#> Time Series:
#> Start = 2004 
#> End = 2015 
#> Frequency = 1 
#>  [1] 720174105 719724239 614310887 617117797 741797161 880065025 928841266
#>  [8] 986589432 924152429 856253649 635471371 790857594
#> 
#> $residuals_fitted$time
#> Time Series:
#> Start = 2004 
#> End = 2015 
#> Frequency = 1 
#>  [1] 2004 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015
#> 
#> $residuals_fitted$line
#> [1] 0
#> 
#> 
#> $compare
#> $compare$resid.variance
#> [1] 1.107099e+16
#> 
#> $compare$variance.coef
#>            ar1        ma1
#> ar1  0.7507992 -0.6832262
#> ma1 -0.6832262  0.7042599
#> 
#> $compare$not.used.obs
#> [1] 0
#> 
#> $compare$used.obs
#> [1] 11
#> 
#> $compare$loglik
#> [1] -217.7045
#> 
#> $compare$aic
#> [1] 441.409
#> 
#> $compare$bic
#> [1] 442.6027
#> 
#> $compare$aicc
#> [1] 444.8376
#> 
#>