Model fit of seasonal time series

Model fit of seasonal time series

ts.seasonal.model(tsdata, x.ord=NULL, tojson=F)

Arguments

tsdata
The input univariate seasonal time series data

x.ord
An integer vector of length 3 specifying the order of the Arima model

tojson
If TRUE the results are returned in json format, default returns a list

Value

A list with the following components:

model.summary:

ts_model The summary model details returned as Arima object for internal use in ts.analysis function

model:

ts_model

arima.order The Arima order

arima.coef A vector of AR, MA and regression coefficients

arima.coef.se The standard error of the coefficients

residuals_fitted:

residuals The residuals of the model (fitted innovations)

fitted The model's fitted values

time the time of tsdata

line The y=0 line

compare:

variance.coef The matrix of the estimated variance of the coefficients

resid.variance The MLE of the innovations variance

not.used.obs The number of not used observations for the fitting

used.obs the number of used observations for the fitting

loglik The maximized log-likelihood (of the differenced data), or the approximation to it used

aic The AIC value corresponding to the log-likelihood

bic The BIC value corresponding to the log-likelihood

aicc The second-order Akaike Information Criterion corresponding to the log-likelihood

Details

Model fit of seasonal time series using arima models of seasonal time series data. The model with the lowest AIC value is selected for forecasts.

See also

ts.analysis, Arima